+++ build_tools/nix/package.nix
@@ -31,7 +31,7 @@ let
 in
 rustPlatform.buildRustPackage {
   pname = "live_compositor";
-  version = "0.2.0";
+  version = "0.3.0";
   src = ../..;
   cargoLock = {
     lockFile = ../../Cargo.lock;